code/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
litellm>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
rich>=13.0.0
pyyaml>=6.0
